clientapp directory restructuring and bug fixes
