--- services/clientapp/src/containers/index.js
@@ -1,25 +1,0 @@
-import Main from "./Main";
-import Home from "./Home";
-import LeftNav from "./LeftNav";
-import Player from "./Player";
-import Auth from "./Auth";
-import Result from "./Result";
-import TopNav from "./TopNav";
-import NowPlaying from "./NowPlaying";
-import PlaylistDisplay from "./PlaylistDisplay";
-import PlaylistManipulator from "./PlaylistManipulator";
-import YourPlaylist from "./YourPlaylist";
-
-export {
-	Main,
-	Home,
-	LeftNav,
-	Player,
-	Auth,
-	TopNav,
-	Result,
-	NowPlaying,
-	PlaylistDisplay,
-	PlaylistManipulator,
-	YourPlaylist
-};
